+++ src/main/java/rw/bk/taxi24/api/service/dto/DriverDTO.java
@@ -14,9 +14,9 @@ public class DriverDTO implements Serializable {
 
     private String name;
 
-    private Float latitude;
+    private Double latitude;
 
-    private Float longitude;
+    private Double longitude;
 
     private DriverStatus status;
 
@@ -36,19 +36,19 @@ public class DriverDTO implements Serializable {
         this.name = name;
     }
 
-    public Float getLatitude() {
+    public Double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(Float latitude) {
+    public void setLatitude(Double latitude) {
         this.latitude = latitude;
     }
 
-    public Float getLongitude() {
+    public Double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(Float longitude) {
+    public void setLongitude(Double longitude) {
         this.longitude = longitude;
     }
